requirements.txt
fastapi
uvicorn
python-multipart
pillow
pypdf2
reportlab
openai
PyMuPDF
openpyxl